requirements.txt
aws-cdk-lib==2.157.0
constructs>=10.0.0,<11.0.0
aws-xray-sdk>=2.14.0
requests>=2.32.3
bs4>=0.0.2
pylint
